+++ demos/supabase-todolist/shared/src/commonMain/kotlin/com/powersync/demos/App.kt
@@ -133,6 +133,7 @@ fun AppContent(
             TodosScreen(
                 modifier = modifier.background(MaterialTheme.colors.background),
                 navController = navController,
+                listId = selectedListId ?: "",
                 items = todoItems,
                 syncStatus = status,
                 inputText = todosInputText,