+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.11.0 (unreleased)
 
 - Update PowerSync core extension to version 0.4.11.
+- Remove the experimental label from Sync Stream APIs.
+- Compose: add `composeSyncStream` helper method to subscribe to Sync Streams in a composition.
 
 ## 1.10.4